refactor: 파일명 및 파일구조 변경
