+++ src/entities/profile/my-community-post-item/ui/community-post-item.tsx
