+++ src/entities/profile/my-community-post-item/ui/index.tsx
