+++ src/entities/profile/registered-product-item/ui/index.tsx
