+++ src/entities/profile/registered-product-item/ui/registered-product-item-style.ts
