+++ src/entities/profile/review-item/ui/review-item-style.ts
