+++ src/entities/user/ui/user-community-post-item/user-community-post-item.tsx
