+++ src/entities/user/ui/user-review-item/user-review-item.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import * as S from './user-review-item-style';
 import Image from 'next/image';
-import { StarRatingResult } from '@entities/star-rating-result/number-to-star';
+import { StarRatingResult } from '@shared/ui/star-rating-result/number-to-star';
 import { formatDateShorter } from '@shared/utils/time';
 
 export const UserReviewItem = () => {
