+++ src/shared/ui/star-rating-result/number-to-star/star-rating-result.tsx
