+++ src/widgets/my-login-info-area/index.tsx
@@ -1,0 +1,1 @@
+export { MyLoginInfoArea } from './my-login-info-area';
