+++ src/widgets/my-login-info-area/my-login-info-area.tsx
@@ -1,14 +1,14 @@
 import { LogoutButton } from '@features/button/logout-button';
 import { RegisterProductButton } from '@features/button/register-product-button';
 
-import * as S from './user-login-info-area-style';
+import * as S from './my-login-info-area-style';
 
-export const UserLoginInfoArea = () => {
+export const MyLoginInfoArea = () => {
   return (
-    <S.UserLoginInfoArea>
+    <S.MyLoginInfoArea>
       <S.LoginInfo>카카오톡으로 로그인 중</S.LoginInfo>
       <LogoutButton />
       <RegisterProductButton />
-    </S.UserLoginInfoArea>
+    </S.MyLoginInfoArea>
   );
 };
