+++ src/widgets/my-transaction-history-list/transaction-history-list-style.ts
