+++ src/widgets/my-wish-list/index.tsx
