+++ src/widgets/my-wish-list/wish-list-style.ts
