+++ src/widgets/my-wish-list/wish-list.tsx
