+++ src/widgets/profile-page-tap-menu-bar/tap-menu-bar-style.ts
