+++ src/widgets/user-community-list/user-community-list.tsx
@@ -1,4 +1,4 @@
-import { UserCommunityPostItem } from '@entities/user-community-post-item';
+import { UserCommunityPostItem } from '@entities/user/ui/user-community-post-item';
 import * as S from './user-community-list-style';
 
 export const UserCommunityList = () => {
