--- src/widgets/user-login-info-area/index.tsx
@@ -1,1 +1,0 @@
-export { UserLoginInfoArea } from './user-login-info-area';
